Create options module
