+++ client/js/localStorage.js
@@ -1,0 +1,13 @@
+"use strict";
+
+module.exports = {
+	set: function(key, value) {
+		try {
+			window.localStorage.setItem(key, value);
+		} catch (e) {
+			// Do nothing. If we end up here, web storage quota exceeded, or user is
+			// in Safari's private browsing where localStorage's setItem is not
+			// available. See http://stackoverflow.com/q/14555347/1935861.
+		}
+	}
+};
