+++ client/js/lounge.js
@@ -15,6 +15,7 @@ const helpers_roundBadgeNumber = require("./libs/handlebars/roundBadgeNumber");
 const slideoutMenu = require("./libs/slideout");
 const templates = require("../views");
 const socket = require("./socket");
+const storage = require("./localStorage");
 const constants = require("./constants");
 
 $(function() {
@@ -39,15 +40,6 @@ $(function() {
 
 	var favicon = $("#favicon");
 
-	function setLocalStorageItem(key, value) {
-		try {
-			window.localStorage.setItem(key, value);
-		} catch (e) {
-			// Do nothing. If we end up here, web storage quota exceeded, or user is
-			// in Safari's private browsing where localStorage's setItem is not
-			// available. See http://stackoverflow.com/q/14555347/1935861.
-		}
-	}
 	socket.on("auth", function(data) {
 		var login = $("#sign-in");
 		var token;
@@ -106,7 +98,7 @@ $(function() {
 		}
 
 		if (data.token && window.localStorage.getItem("token") !== null) {
-			setLocalStorageItem("token", data.token);
+			storage.set("token", data.token);
 		}
 
 		passwordForm
@@ -129,7 +121,7 @@ $(function() {
 		}
 
 		if (data.token && $("#sign-in-remember").is(":checked")) {
-			setLocalStorageItem("token", data.token);
+			storage.set("token", data.token);
 		} else {
 			window.localStorage.removeItem("token");
 		}
@@ -197,7 +189,7 @@ $(function() {
 		var chan = chat.find(target);
 		var template = "msg";
 
-		if (!data.msg.highlight && !data.msg.self && (type === "message" || type === "notice") && highlights.some(function(h) {
+		if (!data.msg.highlight && !data.msg.self && (type === "message" || type === "notice") && options.highlights.some(function(h) {
 			return data.msg.text.toLocaleLowerCase().indexOf(h.toLocaleLowerCase()) > -1;
 		})) {
 			data.msg.highlight = true;
@@ -519,126 +511,9 @@ $(function() {
 
 	socket.on("names", renderChannelUsers);
 
-	var userStyles = $("#user-specified-css");
-	var highlights = [];
-	var options = $.extend({
-		coloredNicks: true,
-		desktopNotifications: false,
-		join: true,
-		links: true,
-		mode: true,
-		motd: false,
-		nick: true,
-		notification: true,
-		notifyAllMessages: false,
-		part: true,
-		quit: true,
-		theme: $("#theme").attr("href").replace(/^themes\/(.*).css$/, "$1"), // Extracts default theme name, set on the server configuration
-		thumbnails: true,
-		userStyles: userStyles.text(),
-	}, JSON.parse(window.localStorage.getItem("settings")));
+	var options = require("./options");
 
 	var windows = $("#windows");
-
-	(function SettingsScope() {
-		var settings = $("#settings");
-
-		for (var i in options) {
-			if (i === "userStyles") {
-				if (!/[?&]nocss/.test(window.location.search)) {
-					$(document.head).find("#user-specified-css").html(options[i]);
-				}
-				settings.find("#user-specified-css-input").val(options[i]);
-			} else if (i === "highlights") {
-				settings.find("input[name=" + i + "]").val(options[i]);
-			} else if (i === "theme") {
-				$("#theme").attr("href", "themes/" + options[i] + ".css");
-				settings.find("select[name=" + i + "]").val(options[i]);
-			} else if (options[i]) {
-				settings.find("input[name=" + i + "]").prop("checked", true);
-			}
-		}
-
-		settings.on("change", "input, select, textarea", function() {
-			var self = $(this);
-			var name = self.attr("name");
-
-			if (self.attr("type") === "checkbox") {
-				options[name] = self.prop("checked");
-			} else {
-				options[name] = self.val();
-			}
-
-			setLocalStorageItem("settings", JSON.stringify(options));
-
-			if ([
-				"join",
-				"mode",
-				"motd",
-				"nick",
-				"part",
-				"quit",
-				"notifyAllMessages",
-			].indexOf(name) !== -1) {
-				chat.toggleClass("hide-" + name, !self.prop("checked"));
-			} else if (name === "coloredNicks") {
-				chat.toggleClass("colored-nicks", self.prop("checked"));
-			} else if (name === "theme") {
-				$("#theme").attr("href", "themes/" + options[name] + ".css");
-			} else if (name === "userStyles") {
-				userStyles.html(options[name]);
-			} else if (name === "highlights") {
-				var highlightString = options[name];
-				highlights = highlightString.split(",").map(function(h) {
-					return h.trim();
-				}).filter(function(h) {
-					// Ensure we don't have empty string in the list of highlights
-					// otherwise, users get notifications for everything
-					return h !== "";
-				});
-			}
-		}).find("input")
-			.trigger("change");
-
-		$("#desktopNotifications").on("change", function() {
-			if ($(this).prop("checked") && Notification.permission !== "granted") {
-				Notification.requestPermission(updateDesktopNotificationStatus);
-			}
-		});
-
-		// Updates the checkbox and warning in settings when the Settings page is
-		// opened or when the checkbox state is changed.
-		// When notifications are not supported, this is never called (because
-		// checkbox state can not be changed).
-		var updateDesktopNotificationStatus = function() {
-			if (Notification.permission === "denied") {
-				desktopNotificationsCheckbox.attr("disabled", true);
-				desktopNotificationsCheckbox.attr("checked", false);
-				warningBlocked.show();
-			} else {
-				if (Notification.permission === "default" && desktopNotificationsCheckbox.prop("checked")) {
-					desktopNotificationsCheckbox.attr("checked", false);
-				}
-				desktopNotificationsCheckbox.attr("disabled", false);
-				warningBlocked.hide();
-			}
-		};
-
-		// If browser does not support notifications, override existing settings and
-		// display proper message in settings.
-		var desktopNotificationsCheckbox = $("#desktopNotifications");
-		var warningUnsupported = $("#warnUnsupportedDesktopNotifications");
-		var warningBlocked = $("#warnBlockedDesktopNotifications");
-		warningBlocked.hide();
-		if (("Notification" in window)) {
-			warningUnsupported.hide();
-			windows.on("show", "#settings", updateDesktopNotificationStatus);
-		} else {
-			options.desktopNotifications = false;
-			desktopNotificationsCheckbox.attr("disabled", true);
-			desktopNotificationsCheckbox.attr("checked", false);
-		}
-	}());
 
 	var viewport = $("#viewport");
 	var sidebarSlide = slideoutMenu(viewport[0], sidebar[0]);
@@ -1227,7 +1102,7 @@ $(function() {
 			}
 		});
 		if (values.user) {
-			setLocalStorageItem("user", values.user);
+			storage.set("user", values.user);
 		}
 		socket.emit(
 			event, values
